funções corretas

+++ pages/api/dados.js
@@ -1,31 +1,29 @@
-import {
-  getDocs,
-  collection,
-  addDoc,
-  deleteDoc,
-  doc,
-} from "firebase/firestore";
-import db from "../../firebase";
+import { getDocs, collection, doc, getDoc, deleteDoc } from "firebase/firestore";
+import db from "@/firebase";
 
 export default async function handler(req, res) {
-  try {
-    const dadosCollection = collection(db, "dados");
+  const dadosCollection = collection(db, "dados");
 
-    if (req.method === "GET") {
-      const { id } = req.query;
+  if (req.method === "GET") {
+    const { id } = req.query;
 
+    try {
       if (id) {
-        const dadosSnapshot = await getDocs(dadosCollection);
-        const dados = dadosSnapshot.docs.map((doc) => ({
-          id: doc.id,
-          ...doc.data(),
-        }));
-        
-        // Aqui ele retorna os dados de acordo com a ID
-        const selectedItem = dados.find((item) => item.id === id);
-        res.status(200).json(selectedItem || {});
+        // Se tiver um ID, busca os dados da ID específica
+        const docRef = doc(dadosCollection, id);
+        const docSnapshot = await getDoc(docRef);
+
+        if (docSnapshot.exists()) {
+          const selectedData = {
+            id: docSnapshot.id,
+            ...docSnapshot.data(),
+          };
+          res.status(200).json(selectedData);
+        } else {
+          res.status(404).json({ error: "Item não encontrado" });
+        }
       } else {
-        // Aqui ele vai retornar todos os dados
+        // Se não tiver ID, busca todos os dados
         const dadosSnapshot = await getDocs(dadosCollection);
         const dados = dadosSnapshot.docs.map((doc) => ({
           id: doc.id,
@@ -33,40 +31,33 @@ export default async function handler(req, res) {
         }));
         res.status(200).json(dados);
       }
-    } else if (req.method === "POST") {
-      const { nome, dataDeEnvio, email, telefone, estado, site } = req.body;
-      const newDocRef = await addDoc(dadosCollection, {
-        nome,
-        dataDeEnvio,
-        email,
-        telefone,
-        estado,
-        site,
-      });
-      res.status(201).json({ id: newDocRef.id });
-    } else if (req.method === "DELETE") {
-      const { id } = req.query;
-
-      if (!id) {
-        res.status(400).json({ error: "ID não fornecida para exclusão" });
-        return;
-      }
-
-      try {
-        const docRef = doc(dadosCollection, id);
-        await deleteDoc(docRef);
-
-        console.log("Documento excluído com sucesso");
-        res.status(200).json({ message: "Documento excluído com sucesso" });
-      } catch (error) {
-        console.error("Erro ao excluir documento:", error);
-        res.status(500).json({ error: "Erro ao excluir documento" });
-      }
-    } else {
-      res.status(405).json({ error: "Método não permitido" });
+    } catch (error) {
+      console.error("Erro ao buscar dados:", error);
+      res.status(500).json({ error: "Erro ao buscar dados." });
     }
-  } catch (error) {
-    console.error("Erro ao lidar com a requisição:", error);
-    res.status(500).json({ error: "Erro ao lidar com a requisição." });
+  } else if (req.method === "DELETE") {
+    const { id } = req.query;
+  
+    if (!id) {
+      res.status(400).json({ error: "ID não fornecida para exclusão" });
+      return;
+    }
+  
+    try {
+      const docRef = doc(dadosCollection, id);
+      const docSnapshot = await getDoc(docRef);
+  
+      if (docSnapshot.exists()) {
+        await deleteDoc(docRef);
+        res.status(200).json({ message: "Documento excluído com sucesso" });
+      } else {
+        res.status(404).json({ error: "Item não encontrado" });
+      }
+    } catch (error) {
+      console.error("Erro ao excluir documento:", error);
+      res.status(500).json({ error: "Erro ao excluir documento" });
+    }
+  } else {
+    res.status(405).json({ error: "Método não permitido" });
   }
 }
